+++ nafas/params.py
@@ -99,7 +99,7 @@ STANDARD_MENU = {
         15: "Energizing",
         16: "Box",
         17: "Coherent",
-        18: "Fire Breath",
+        18: "Fire",
         19: "Breath Retention",
         20: "Swooning Breath"},
     "level": {
@@ -153,11 +153,11 @@ HARMONY_MEDIUM = {"ratio": [1, 4, 2, 1], "unit": 3, "pre": 3, "cycle": 18}
 
 HARMONY_ADVANCED = {"ratio": [1, 5, 2, 1], "unit": 3, "pre": 3, "cycle": 16}
 
-FIRE_BREATH_BEGINNER = {"ratio": [3, 0, 5, 12], "unit": 1, "pre": 3, "cycle": 10}
+FIRE_BEGINNER = {"ratio": [3, 0, 5, 12], "unit": 1, "pre": 3, "cycle": 10}
 
-FIRE_BREATH_MEDIUM = {"ratio": [3, 0, 5, 12], "unit": 1, "pre": 3, "cycle": 20}
+FIRE_MEDIUM = {"ratio": [3, 0, 5, 12], "unit": 1, "pre": 3, "cycle": 20}
 
-FIRE_BREATH_ADVANCED = {"ratio": [3, 0, 5, 12], "unit": 1, "pre": 3, "cycle": 30}
+FIRE_ADVANCED = {"ratio": [3, 0, 5, 12], "unit": 1, "pre": 3, "cycle": 30}
 
 BREATH_RETENTION_BEGINNER = {"ratio": [2, 0, 2, 12], "unit": 1, "pre": 3, "cycle": 5}
 
@@ -531,9 +531,9 @@ PROGRAMS = {"Clear Mind": {"Beginner": CLEAR_MIND_BEGINNER,
             "Coherent": {"Beginner": COHERENT_BEGINNER,
                          "Medium": COHERENT_MEDIUM,
                          "Advanced": COHERENT_ADVANCED},
-            "Fire Breath": {"Beginner": FIRE_BREATH_BEGINNER,
-                            "Medium": FIRE_BREATH_MEDIUM,
-                            "Advanced": FIRE_BREATH_ADVANCED},
+            "Fire": {"Beginner": FIRE_BEGINNER,
+                            "Medium": FIRE_MEDIUM,
+                            "Advanced": FIRE_ADVANCED},
             "Breath Retention": {"Beginner": BREATH_RETENTION_BEGINNER,
                                  "Medium": BREATH_RETENTION_MEDIUM,
                                  "Advanced": BREATH_RETENTION_ADVANCED},