+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - `--generate-config` argument
+- `--color` argument
 ### Changed
 - `README.md` modified
 - `Python 3.14` added to `test.yml`