+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `--generate-config` argument
 - `--color` argument
+- `--bg-color` argument
 ### Changed
 - `README.md` modified
 - `Python 3.14` added to `test.yml`