+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `README.md` modified
 - `Python 3.14` added to `test.yml`
 - Test system modified
+- CLI functions modified
 ## [1.4] - 2025-07-29
 ### Added
 - `ca1` speaker audio files