+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.5] - 2026-03-23
 ### Added
 - Fire program
 - Retention program
@@ -188,7 +189,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Anti-Appetite program 
 - Cigarette Replace program
 
-[Unreleased]: https://github.com/sepandhaghighi/nafas/compare/v1.4...dev
+[Unreleased]: https://github.com/sepandhaghighi/nafas/compare/v1.5...dev
+[1.5]: https://github.com/sepandhaghighi/nafas/compare/v1.4...v1.5
 [1.4]: https://github.com/sepandhaghighi/nafas/compare/v1.3...v1.4
 [1.3]: https://github.com/sepandhaghighi/nafas/compare/v1.2...v1.3
 [1.2]: https://github.com/sepandhaghighi/nafas/compare/v1.1...v1.2